use trigger rule simulation engine to validate resource consumption issues for Cats.daml example
