+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/CatTriggerResourceUsageTest.scala
@@ -1,0 +1,47 @@
+// Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.lf
+package engine.trigger
+
+import engine.trigger.test.AbstractTriggerTest
+
+import com.daml.ledger.api.testing.utils.SuiteResourceManagementAroundAll
+import com.daml.ledger.runner.common.Config
+import com.daml.ledger.sandbox.SandboxOnXForTest.{ApiServerConfig, singleParticipant}
+import com.daml.platform.services.time.TimeProviderType
+import org.scalatest.{Inside, TryValues}
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AsyncWordSpec
+
+class CatTriggerResourceUsageTest
+    extends AsyncWordSpec
+    with AbstractTriggerTest
+    with Matchers
+    with Inside
+    with SuiteResourceManagementAroundAll
+    with TryValues
+    with CatGenerators {
+
+  import TriggerRuleSimulationLib._
+
+  override protected def config: Config = super.config.copy(
+    participants = singleParticipant(
+      ApiServerConfig.copy(
+        timeProviderType = TimeProviderType.Static
+      )
+    )
+  )
+
+  override protected def triggerRunnerConfiguration: TriggerRunnerConfig =
+    super.triggerRunnerConfiguration.copy(hardLimit =
+      super.triggerRunnerConfiguration.hardLimit
+        .copy(allowTriggerTimeouts = true, allowInFlightCommandOverflows = true)
+    )
+
+  "Trigger rule simulation" should {
+    "identify that number of submissions is dependent on ACS size" in {
+      succeed // TODO:
+    }
+  }
+}
